order-service/requirements.txt
fastapi
uvicorn
psycopg2-binary
requests
prometheus-fastapi-instrumentator